Refactor knowpro test sample
